tests/wpt/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/protocol.https.html

<!DOCTYPE html>
<meta charset='utf-8'>
<title>protocol handlers</title>

<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>

<noscript><p>Enable JavaScript and reload.</p></noscript>

<p><strong>Note:</strong> If your browser limits the number of handler
registration requests on a page, you might need to disable or significantly
increase that limit for the tests below to run.</p>

<script type='text/javascript'>
test(() => {
  assert_idl_attribute(navigator, 'registerProtocolHandler');
}, 'the registerProtocolHandler method should exist on the navigator object');

test(() => {
  assert_idl_attribute(navigator, 'unregisterProtocolHandler');
}, 'the unregisterProtocolHandler method should exist on the navigator object');

/* URL argument */
const valid_urls = [
  '%s',
  'foo/%s',
  location.href + '/%s',
  location.href + '#%s',
  location.href + '?foo=%s',
  location.href + '?foo=%s&bar',
  location.href + '/%s/bar/baz/',
  location.href + '/%s/bar/baz/?foo=1337&bar#baz',
  location.href + '/%s/foo/%s/',
];
for (const url of valid_urls) {
  test(() => {
    navigator.registerProtocolHandler('tel', url, 'foo');
  }, 'registerProtocolHandler: Valid URL "' + url + '" should work.');

  test(() => {
    navigator.unregisterProtocolHandler('tel', url);
  }, 'unregisterProtocolHandler: Valid URL "' + url + '" should work.');
}

const invalid_urls1 = [
  '',
  '%S',
  location.href + '',
  location.href + '/%',
  location.href + '/%a',
  'http://example.com',
  'http://[v8.:::]//url=%s',
];

for (const url of invalid_urls1) {
  test(() => {
    assert_throws_dom('SYNTAX_ERR', () => { navigator.registerProtocolHandler('mailto', url, 'foo'); });
  }, 'registerProtocolHandler: Invalid URL "' + url + '" should throw SYNTAX_ERR.');

  test(() => {
    assert_throws_dom('SYNTAX_ERR', () => { navigator.unregisterProtocolHandler('mailto', url); });
  }, 'unregisterProtocolHandler: Invalid URL "' + url + '" should throw SYNTAX_ERR.');
}

const invalid_urls2 = [
  'http://%s.com',
  'http://%s.example.com',
  'http://example.com/%s',
  'https://example.com/%s',
  'http://foobar.example.com/%s',
  'mailto:%s@example.com',
  'mailto:%s',
  `ftp://${location.host}/%s`,
  `chrome://${location.host}/%s`,
  `foo://${location.host}/%s`,
];
for (const url of invalid_urls2) {
  test(() => {
    assert_throws_dom('SECURITY_ERR', () => { navigator.registerProtocolHandler('mailto', url, 'foo'); });
  }, 'registerProtocolHandler: Invalid URL "' + url + '" should throw SECURITY_ERR.');

  test(() => {
    assert_throws_dom('SECURITY_ERR', () => { navigator.unregisterProtocolHandler('mailto', url); });
  }, 'unregisterProtocolHandler: Invalid URL "' + url + '" should throw SECURITY_ERR.');
}

/* Protocol argument */

/* Overriding any of the following protocols must never be allowed. That would
 * break the browser. */
const denylist = [
  'about',
  'attachment',
  'blob',
  'chrome',
  'cid',
  'data',
  'file',
  'ftp',
  'http',
  'https',
  'javascript',
  'livescript',
  'mid',
  'mocha',
  'moz-icon',
  'opera',
  'operamail',
  'res',
  'resource',
  'shttp',
  'tcl',
  'vbscript',
  'view-source',
  'ws',
  'wss',
  'wyciwyg',
  /* other invalid schemes */
  'unrecognized',
  'mаilto', /* a cyrillic "а" */
  'mailto:',
  'mailto://',
  'mailto' + String.fromCharCode(0),
  'mailtoo' + String.fromCharCode(8),
  'mailto' + String.fromCharCode(10),
  'http://',
  'ssh:/',
  'magnet:+',
  'tel:sip',
  'foo',
  'fweb+oo',
  /* web+ prefixed schemes must be followed by 1+ ascii alphas */
  'web+',
  'web+1',
  'web+dots.are.forbidden',
  'web+dashes-are-forbidden',
  'web+digits123areforbidden',
  'web+UpperCasesAreForbidden',
];
for (const scheme of denylist) {
  test(() => {
    assert_throws_dom('SECURITY_ERR', () => { navigator.registerProtocolHandler(scheme, location.href + '/%s', 'foo'); });
  }, 'registerProtocolHandler: Attempting to override the "' + scheme + '" protocol should throw SECURITY_ERR.');

  test(() => {
    assert_throws_dom('SECURITY_ERR', () => { navigator.unregisterProtocolHandler(scheme, location.href + '/%s'); });
  }, 'unregisterProtocolHandler: Attempting to override the "' + scheme + '" protocol should throw SECURITY_ERR.');
}

/* The following protocols must be possible to override.
 * We're just testing that the call goes through here. Whether or not they
 * actually work as handlers is covered by the interactive tests. */

const safelist = [
  /* safelisted schemes listed in
   * https://html.spec.whatwg.org/multipage/system-state.html#safelisted-scheme */
  'bitcoin',
  'geo',
  'im',
  'irc',
  'ircs',
  'magnet',
  'mailto',
  'mms',
  'news',
  'nntp',
  'openpgp4fpr',
  'sip',
  'sms',
  'smsto',
  'ssh',
  'tel',
  'urn',
  'webcal',
  'wtai',
  'xmpp',
  /* other valid schemes */
  'BitcoIn',
  'Irc',
  'MagneT',
  'SmsTo',
  'TEL',
  'teL',
  'WebCAL',
  'WTAI',
  'web+myprotocol',
];
for (const scheme of safelist) {
  test(() => {
    navigator.registerProtocolHandler(scheme, location.href + '/%s', "foo");
  }, 'registerProtocolHandler: overriding the "' + scheme + '" protocol should work');

  test(() => {
    navigator.unregisterProtocolHandler(scheme, location.href + '/%s');
  }, 'unregisterProtocolHandler: overriding the "' + scheme + '" protocol should work');
}
</script>
